+++ tests/Unit/CosmosDB.utils.Tests.ps1
@@ -163,7 +163,7 @@ console.log("done");
             { Get-Command -Name Convert-CosmosDbSecureStringToString -ErrorAction Stop } | Should -Not -Throw
         }
 
-        Context 'When called with valid secure string' {
+        Context 'When called with valid SecureString' {
             $script:result = $null
 
             It 'Should not throw exception' {
@@ -172,6 +172,22 @@ console.log("done");
                 }
 
                 { $script:result = Convert-CosmosDbSecureStringToString @convertCosmosDbSecureStringToStringParameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result = $script:testKey
+            }
+        }
+
+        Context 'When called with a valid SecureString via pipeline variable' {
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $convertCosmosDbSecureStringToStringParameters = @{
+                    SecureString = $script:testKeySecureString
+                }
+                Write-Verbose -Message "Type: $($script:testKeySecureString.GetType().FullName)" -Verbose
+                { $script:result = $convertCosmosDbSecureStringToStringParameters.SecureString | Convert-CosmosDbSecureStringToString } | Should -Not -Throw
             }
 
             It 'Should return expected result' {
@@ -2248,13 +2264,20 @@ console.log("done");
                 { $script:result = Get-CosmosDbEntraIdToken } | Should -Not -Throw
             }
 
-            It 'Should return secure string containing token' {
+            It 'Should return a secure string' {
+                $script:result | Should -BeOfType [System.Security.SecureString]
+            }
+
+            It 'Should return secure string containing the expected token' {
                 $script:result | Convert-CosmosDbSecureStringToString | Should -Be $script:testEntraIdToken
             }
 
             It 'Should call expected mocks' {
                 Assert-MockCalled -CommandName Get-AzAccessToken -Exactly -Times 1 `
-                    -ParameterFilter { $ResourceUrl -eq 'https://cosmos.azure.com' }
+                    -ParameterFilter {
+                        $AsSecureString -and
+                        $ResourceUrl -eq 'https://cosmos.azure.com'
+                    }
             }
         }
 
@@ -2274,7 +2297,11 @@ console.log("done");
                 { $script:result = Get-CosmosDbEntraIdToken @getCosmosDbEntraIdTokenParameters } | Should -Not -Throw
             }
 
-            It 'Should return secure string containing token' {
+            It 'Should return a secure string' {
+                $script:result | Should -BeOfType [System.Security.SecureString]
+            }
+
+            It 'Should return secure string containing the expected token' {
                 $script:result | Convert-CosmosDbSecureStringToString | Should -Be $script:testEntraIdToken
             }
 
@@ -2300,7 +2327,11 @@ console.log("done");
                 { $script:result = Get-CosmosDbEntraIdToken @getCosmosDbEntraIdTokenParameters } | Should -Not -Throw
             }
 
-            It 'Should return secure string containing token' {
+            It 'Should return a secure string' {
+                $script:result | Should -BeOfType [System.Security.SecureString]
+            }
+
+            It 'Should return secure string containing the expected token' {
                 $script:result | Convert-CosmosDbSecureStringToString | Should -Be $script:testEntraIdToken
             }
 
@@ -2317,8 +2348,13 @@ console.log("done");
                 return $null
             }
 
-            It 'Should throw exception' {
-                { $script:result = Get-CosmosDbEntraIdToken } | Should -Not -Throw
+            $errorRecord = Get-InvalidOperationRecord `
+                -Message ($LocalizedData.ErrorGettingEntraIdToken -f 'https://cosmos.azure.com')
+
+            It 'Should throw expected exception' {
+                {
+                    $script:result = Get-CosmosDbEntraIdToken
+                } | Should -Throw $errorRecord
             }
 
             It 'Should return null' {